move window-frame-listener
